cleaner tweak

+++ manipulator/cleaner.py
@@ -6,7 +6,7 @@ import json
 
 
 if __name__ == '__main__':
-    ts = common.get_all_good_traces()
+    ts = common.get_all_traces()
     groups = {}
     for t in ts:
         key = t['prefix']+str(t['id'])+t['solver']
@@ -20,12 +20,13 @@ if __name__ == '__main__':
         rubbish = l[2:]
         for t in rubbish:
             meta = {}
-            with io.open(t['meta_fname'], 'r') as f:
-                meta = json.loads(f.read())
-            if not 'rubbish' in meta:
-                meta['rubbish'] = True
-                with io.open(t['meta_fname'], 'w') as f:
-                    f.write(json.dumps(meta))
+            if os.path.isfile(t['meta_fname']):
+                with io.open(t['meta_fname'], 'r') as f:
+                    meta = json.loads(f.read())
+                if not 'rubbish' in meta:
+                    meta['rubbish'] = True
+                    with io.open(t['meta_fname'], 'w') as f:
+                        f.write(json.dumps(meta))
             if os.path.isfile(t['fname']):
                 os.remove(t['fname'])
                 print('%s was marked as rubbish and removed' % t['fname'])
